✨ feat - Refreshes the wallet delete page.
