+++ Source/Mobile/Finance/Finance.csproj.user
@@ -35,7 +35,7 @@
     <MauiXaml Update="Pages\DangerZone\DangerZonePage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Pages\DangerZone\Delete\DeleteWalletPage.xaml">
+    <MauiXaml Update="Pages\Walleting\DeleteWalletPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
     <MauiXaml Update="Pages\Walleting\CreateWalletPage.xaml">
